migrate to stdint.h, but remove unneccesary _t at the end (int32_t -> int32)

+++ src/91_nineone/include/libc/stdio.h
@@ -1,5 +1,5 @@
 #pragma once
 
 int putchar(int ic);
-bool print(const char* data, size_t length);
+bool print(const char* data, uint64 length);
 int printf(const char* __restrict__ format, ...);
